Move team config and plan endpoints to v3 router
Refactored API endpoints for team configuration and plan management from app_kernel.py to v3/api/router.py, consolidating related logic under the new app_v3 router with '/api/v3' prefix. Removed redundant imports and endpoint definitions from app_kernel.py, improving code organization and separation of concerns.
